divided main.dart file into four separate files.
