+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
+  rflutter_alert: ^2.0.2
 
 dev_dependencies:
   flutter_test:
